+++ core/src/workflows/compliance_updater/events.py
@@ -8,9 +8,6 @@ class ComplianceBasicInformation(Event):
     prowler_provider: str = Field(
         description="Prowler provider to use for the check creation"
     )
-    service: str = Field(
-        description="Service of the provider to which the check is related"
-    )
     compliance_data: dict = Field(description="Compliance data to update")
