+++ core/src/workflows/compliance_updater/workflow.py
@@ -1,13 +1,14 @@
-from llama_index.core import Settings
-from llama_index.core.workflow import Context, StartEvent, StopEvent, Workflow, step
+from llama_index.core.workflow import StartEvent, StopEvent, Workflow, step
 from llama_index.core.workflow.retry_policy import ConstantDelayRetryPolicy
 from loguru import logger
 
 from core.src.rag.vector_store import CheckMetadataVectorStore
-from core.src.utils.model_chooser import llm_chooser
 from core.src.workflows.compliance_updater.events import (
     ComplianceBasicInformation,
     ComplianceDataResult,
+)
+from core.src.workflows.compliance_updater.utils.compliance_validator import (
+    is_valid_prowler_compliance,
 )
 
 
@@ -16,59 +17,39 @@ class ComplianceUpdaterWorkflow(Workflow):
 
     @step(retry_policy=ConstantDelayRetryPolicy(delay=10, maximum_attempts=3))
     async def workflow_setup(
-        self, ctx: Context, start_event: StartEvent
+        self, start_event: StartEvent
     ) -> ComplianceBasicInformation | StopEvent:
         """Setup the workflow and sanitize the user input for next steps.
         Args:
-            ctx: Workflow context.
             user_query: User input to start the workflow.
         """
         logger.info("Initializing...")
         try:
             compliance_data = start_event.get("compliance_data", "")
 
-            if compliance_data:
-                await ctx.set("compliance_data", compliance_data)
-
-                Settings.llm = llm_chooser(
-                    model_provider=start_event.get("model_provider", ""),
-                    model_reference=start_event.get("model_reference", ""),
-                    api_key=start_event.get("api_key", ""),
-                )
-
-                await ctx.set("model_provider", start_event.get("model_provider"))
-                await ctx.set("model_reference", start_event.get("model_reference"))
-
+            if is_valid_prowler_compliance(compliance_data):
                 return ComplianceBasicInformation(
                     prowler_provider=compliance_data.get("Provider", "").lower(),
-                    service="",
                     compliance_data=compliance_data,
                 )
-
             else:
-                raise ValueError("The provided compliance data is empty.")
+                raise ValueError("Invalid Prowler compliance data format.")
 
-        except ValueError as e:
-            logger.error(str(e))
-            return StopEvent()
         except Exception as e:
             logger.exception(e)
-            return StopEvent()
+            return StopEvent(result=str(e))
 
     @step(retry_policy=ConstantDelayRetryPolicy(delay=5, maximum_attempts=3))
     async def get_compliance_checks(
-        self, ctx: Context, compliance_basic_info: ComplianceBasicInformation
+        self, compliance_basic_info: ComplianceBasicInformation
     ) -> ComplianceDataResult:
-        """Update the compliance data with the relevant checks.
+        """Get the relevant checks for the compliance data.
         Args:
-            ctx: Workflow context.
             check_basic_info: Basic information extracted from the user query to update the compliance data.
         """
-        logger.info("Updating compliance data...")
+        logger.info("Retrieving relevant checks...")
         try:
             check_metadata_vector_store = CheckMetadataVectorStore()
-
-            await ctx.set("check_metadata_vector_store", check_metadata_vector_store)
 
             output_data = []
             for requirement in compliance_basic_info.compliance_data["Requirements"]:
@@ -94,22 +75,16 @@ class ComplianceUpdaterWorkflow(Workflow):
                 original_compliance_data=compliance_basic_info.compliance_data,
                 updated_compliance_data=output_data,
             )
-
-        except ValueError as e:
-            logger.error(str(e))
-            return StopEvent()
-
         except Exception as e:
             logger.exception(e)
-            return StopEvent()
+            return StopEvent(result=str(e))
 
     @step(retry_policy=ConstantDelayRetryPolicy(delay=5, maximum_attempts=3))
     async def update_compliance(
-        self, ctx: Context, compliance_data: ComplianceDataResult
+        self, compliance_data: ComplianceDataResult
     ) -> StopEvent:
         """Update the compliance data with the relevant checks.
         Args:
-            ctx: Workflow context.
             compliance_data: Compliance data with the relevant checks.
         """
         logger.info("Updating compliance data...")
@@ -125,10 +100,6 @@ class ComplianceUpdaterWorkflow(Workflow):
                         break
 
             return StopEvent(result=compliance_data.original_compliance_data)
-        except ValueError as e:
-            logger.error(str(e))
-            return StopEvent()
-
         except Exception as e:
             logger.exception(e)
-            return StopEvent()
+            return StopEvent(result=str(e))
